Add bytes label component
