+++ gui/src/pages/Process.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
   import * as router from 'svelte-spa-router';
+  import sparkline from '@fnando/sparkline';
   import Layout from '../components/Layout.svelte';
   import { api } from '../lib/api';
   import { onDestroy, onMount } from 'svelte';
@@ -10,7 +11,7 @@
   } from '../lib/process/store';
   import type { RemoteData } from '../lib/api';
   import type { ComponentDetails } from '../lib/process/types';
-  import sparkline from '@fnando/sparkline';
+  import BytesLabel from '../components/BytesLabel.svelte';
   export let params = { workspace: '', process: '' };
 
   const workspaceId = params.workspace;
@@ -58,41 +59,6 @@
     clearInterval(refreshInterval);
     unsubscribeProcesses();
   });
-
-  // From https://stackoverflow.com/a/14919494
-  /**
-   * Format bytes as human-readable text.
-   *
-   * @param bytes Number of bytes.
-   * @param si True to use metric (SI) units, aka powers of 1000. False to use
-   *           binary (IEC), aka powers of 1024.
-   * @param dp Number of decimal places to display.
-   *
-   * @return Formatted string.
-   */
-  function humanFileSize(bytes: number, si = false, dp = 1): string {
-    const thresh = si ? 1000 : 1024;
-
-    if (Math.abs(bytes) < thresh) {
-      return bytes + ' B';
-    }
-
-    const units = si
-      ? ['kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
-      : ['KiB', 'MiB', 'GiB', 'TiB', 'PiB', 'EiB', 'ZiB', 'YiB'];
-    let u = -1;
-    const r = 10 ** dp;
-
-    do {
-      bytes /= thresh;
-      ++u;
-    } while (
-      Math.round(Math.abs(bytes) * r) / r >= thresh &&
-      u < units.length - 1
-    );
-
-    return bytes.toFixed(dp) + ' ' + units[u];
-  }
 </script>
 
 <Layout>
@@ -132,7 +98,7 @@
             </tr>
             <tr>
               <td>Resident Memory</td>
-              <td>{humanFileSize(process.status.residentMemory)}</td>
+              <td><BytesLabel value={process.status.residentMemory} /></td>
             </tr>
             <tr>
               <td>Started at</td>
